# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Preuvely API Documentation'
  description: 'API for Preuvely - Trusted Store Reviews Platform for Algeria'
  version: 1.0.0
  contact:
    name: 'Preuvely Support'
    email: support@preuvely.dz
servers:
  -
    url: 'http://localhost'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Obtain your token by calling POST `/api/v1/auth/login` or POST `/api/v1/auth/register`. Include the token in the Authorization header as `Bearer {token}`.'
security:
  -
    default: []
paths:
...
  /api/v1/auth/email/resend:
    post:
      summary: 'Resend verification email'
      operationId: resendVerificationEmail
      description: 'Resend the email verification notification.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Verification email resent successfully'
                properties:
                  message:
                    type: string
                    example: 'Verification email resent successfully'
        400:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Email already verified'
                properties:
                  message:
                    type: string
                    example: 'Email already verified'
      tags:
        - Authentication
...